+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "python.linting.pylintEnabled": true,
+    "python.linting.enabled": true,
+    "python.pythonPath": "~/Virtualenvs/Python3/omni/bin/python",
+}
