+++ test_bookScrapy.py
@@ -1,24 +1,24 @@
 import pandas as pd
-import bookScrap
+import bookScrapy
 
 from pandas.io import sql
 from sqlalchemy import create_engine
 
 
-driver = bookScrap.driver
+driver = bookScrapy.driver
 
 def test_driver_init():
-    driver = bookScrap.driver_init()
+    driver = bookScrapy.driver_init()
     driver.get('http://books.toscrape.com/index.html')
     title = "All products | Books to Scrape - Sandbox"
     assert title == driver.title
 
 def test_scrapy_category_lists1(driver=driver):
-    category_links, category_list = bookScrap.scrapy_category_lists(driver)
+    category_links, category_list = bookScrapy.scrapy_category_lists(driver)
     assert category_links != None and category_list != None
 
 def test_scrapy_category_lists2(driver=driver):
-    category_links, category_list = bookScrap.scrapy_category_lists(driver)
+    category_links, category_list = bookScrapy.scrapy_category_lists(driver)
 
     category_from_link = [link.split('/')[6].split('_')[0].replace('-', ' ').title() 
                      for link in category_links]
@@ -30,18 +30,18 @@ def test_scrapy_category_lists2(driver=driver):
 
 
 def test_scrapy_rows(driver=driver):
-    assert bookScrap.scrapy_rows(driver) != None
+    assert bookScrapy.scrapy_rows(driver) != None
 
 def test_scrapy_rows2(driver=driver):
     driver.get('http://books.toscrape.com/index.html')
-    rows = bookScrap.scrapy_rows(driver)
+    rows = bookScrapy.scrapy_rows(driver)
     title = rows[0].find_elements_by_xpath('//h3//a')[0].get_attribute("title") 
     assert title == 'A Light in the Attic'
 
 def test_extract_data(driver=driver):
     driver.get('http://books.toscrape.com/index.html')
-    rows = bookScrap.scrapy_rows(driver)
-    data = bookScrap.extract_data(rows, 'All products')
+    rows = bookScrapy.scrapy_rows(driver)
+    data = bookScrapy.extract_data(rows, 'All products')
 
     final_data = [len(data.get('title')),len(data.get('price')),
                   len(data.get('rating')),len(data.get('in_stock')),
@@ -61,17 +61,17 @@ def test_to_postgress():
             }
 
     sql.execute('DROP TABLE IF EXISTS book_club_basic.teste', engine)
-    bookScrap.to_postgres(data,engine, schema='book_club_basic', table='teste')
+    bookScrapy.to_postgres(data,engine, schema='book_club_basic', table='teste')
     df = pd.DataFrame(sql.execute('Select * From book_club_basic.teste', engine),        
     columns = ['title','price','rating','in_stock','category'])
     assert df.shape == (10,5)
 
 def test_next_page_extration1(driver=driver):
     driver.get('http://books.toscrape.com/index.html')
-    next_page = bookScrap.extract_next_page(driver)
+    next_page = bookScrapy.extract_next_page(driver)
     assert next_page == 'http://books.toscrape.com/catalogue/page-2.html'
 
 def test_next_page_extration2(driver=driver):
     driver.get('http://books.toscrape.com/catalogue/page-50.html')
-    next_page = bookScrap.extract_next_page(driver)
+    next_page = bookScrapy.extract_next_page(driver)
     assert next_page == None
